requirements.txt
# ── Core LLM / LangChain ──────────────────────────────────────
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
openai>=1.40.0
anthropic>=0.35.0

# ── Vector Stores ─────────────────────────────────────────────
chromadb>=0.5.0
faiss-cpu>=1.8.0

# ── Embeddings / Reranking ────────────────────────────────────
sentence-transformers>=3.0.0

# ── Web Retrieval ─────────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=6.0.0

# ── Document Loading ──────────────────────────────────────────
pypdf>=4.0.0
python-docx>=1.1.0

# ── App ───────────────────────────────────────────────────────
streamlit>=1.38.0

# ── Config / Utils ────────────────────────────────────────────
pyyaml>=6.0
python-dotenv>=1.0.0

# ── Evaluation ────────────────────────────────────────────────
ragas>=0.1.0
datasets>=2.20.0

# ── Development / Testing ─────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.4.0
